Setup a working directory and add option to auto-submit the jobs
The config files, job scripts, and initial netCDF file are all now
created in a working directory. The working directory defaults to `work`
but can be set with the `-w/--working-dir` option.

Another option, `-r/--run` was added to submit the set of job scripts to
the queue. Because these job scripts are sequential in nature, the
following job scripts will be held in the queue until the current script
finishes.

The `run.py` script was renamed to `spinup.py` because `run.py --run`
doesn't make any sense and `spinup.py` better described what the script
does.