State of initMIP spinup using CISM-Albany at submission deadline
I was never able to get the constant or dsmb test to run completely.
Using the flux correction and no advance mask, I was able to get about
35 years of evolution.

The spinup python script works, but the run one doesn't. Most of the run
stuff was manually performed using the scattering of bash scripts around
in here. See the work/submit directory and you should be able to follow
the workflow.

This will be improved into a working repo at some point. Hopefully.

tl;dr: not working. Epic fail.